Move KryoHelper to use it for different processes communication
