+++ utbot-instrumentation/build.gradle.kts
@@ -49,9 +49,6 @@ dependencies {
 
     implementation("org.ow2.asm:asm:$asmVersion")
     implementation("org.ow2.asm:asm-commons:$asmVersion")
-    implementation("com.esotericsoftware.kryo:kryo5:$kryoVersion")
-    // this is necessary for serialization of some collections
-    implementation("de.javakaffee:kryo-serializers:$kryoSerializersVersion")
     implementation("io.github.microutils:kotlin-logging:$kotlinLoggingVersion")
 
     implementation("com.jetbrains.rd:rd-framework:$rdVersion")
